ログイン情報をauth package内の関数で取得するように変更
